[pattern-core] phase 2 close: zero warnings on cargo check and clippy (AC1.1)
Fix pre-existing clippy lints across pattern-db, pattern-auth (auto-fix),
stage broken integration tests + examples with fate markers, drop the
legacy `messages` test_helpers module that referenced the staged-away
messages/ module and SnowflakePosition. Crate-level #[allow]s suppress
pre-existing style lints in feature-gated export/ + legacy memory/document.rs
that are orthogonal to Phase 2 scope and will be addressed in Phase 3/4.

+++ crates/pattern_core/src/utils.rs
@@ -204,7 +204,7 @@ pub fn get_next_message_position_sync() -> SnowflakePosition {
             IdGenStatus::Pending { yield_for } => {
                 // If yield_for is 0, we're at the sequence limit but still in the same millisecond.
                 // Wait at least 1ms to roll over to the next millisecond and reset the sequence.
-                let wait_ms = yield_for.max(1) as u64;
+                let wait_ms = yield_for.max(1);
                 std::thread::sleep(std::time::Duration::from_millis(wait_ms));
                 // Loop will retry after the wait
             }
